todo/requirements.txt
asgiref==3.5.2
Django==3.2.8
django-cors-headers==3.13.0
django-filter==22.1
djangorestframework==3.13.1
djangorestframework-camel-case==1.3.0
Markdown==3.4.1
pytz==2022.2.1
sqlparse==0.4.2
djangorestframework-simplejwt==5.2.1
requests==2.28.1
drf_yasg==1.21.4
